aiogram-bot-template-master/requirements.txt
aiogram<3.0
environs~=8.0.0
Pillow~=9.0.1
opencv-python~=4.5.5.64
face-recognition=1.3.0
stegano~=0.10.2
requests~=2.27.1
deepface~=0.0.74
translate~=3.6.1